requirements.txt
Django==3.1.7
django-adminlte-ui==1.5.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-jwt==0.3.2
django-treebeard==4.3
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.4
graphql-relay==3.1.0
requests==2.25.1
celery==5.1.0
redis==3.5.3
ippanel==1.1.0
graphene-file-upload==1.3.0
Pillow==8.1.0
zeep==4.0.0
aiohttp==3.7.4.post0